get voting incentives using useQuery (better error management)
