+++ packages/lib/modules/vebal/vote/Votes/VotesContainer.tsx
@@ -1,7 +1,6 @@
 import { GetVeBalVotingListDocument } from '@repo/lib/shared/services/api/generated/graphql'
 import { mins } from '@repo/lib/shared/utils/time'
 import { getApolloServerClient } from '@repo/lib/shared/services/api/apollo-server.client'
-import { parseError } from '@repo/lib/shared/utils/errors'
 import { VotesProvider } from '@repo/lib/modules/vebal/vote/Votes/VotesProvider'
 import { VoteListLayout } from '@repo/lib/modules/vebal/vote/VoteList/VoteListLayout'
 import { MyVotesLayout } from '@repo/lib/modules/vebal/vote/Votes/MyVotes/MyVotesLayout'
@@ -9,7 +8,6 @@ import { VoteListProvider } from '@repo/lib/modules/vebal/vote/VoteList/VoteList
 import { VStack } from '@chakra-ui/react'
 import { MyVotesProvider } from '@repo/lib/modules/vebal/vote/Votes/MyVotes/MyVotesProvider'
 import { TransactionStateProvider } from '@repo/lib/modules/transactions/transaction-steps/TransactionStateProvider'
-import { getHiddenHandVotingIncentivesEither } from '@repo/lib/shared/services/hidden-hand/getHiddenHandVotingIncentives'
 import { VotesIntroductionLayout } from '@repo/lib/modules/vebal/vote/Votes/VotesIntroduction/VotesIntroductionLayout'
 
 export async function VotesContainer() {
@@ -27,16 +25,10 @@ export async function VotesContainer() {
     })
   }
 
-  const [{ data: voteListQueryData }, [votingIncentives, votingIncentivesError]] =
-    await Promise.all([getVebalVotingList(), getHiddenHandVotingIncentivesEither()])
+  const { data: voteListQueryData } = await getVebalVotingList()
 
   return (
-    <VotesProvider
-      data={voteListQueryData}
-      votingIncentives={votingIncentives}
-      votingIncentivesErrorMessage={parseError(votingIncentivesError)}
-      votingIncentivesLoading={false} /* RSC (SSR) mode, no loading needed */
-    >
+    <VotesProvider data={voteListQueryData}>
       <VStack spacing="3xl" w="full">
         <VotesIntroductionLayout />
